+++ src/ThrowReporter.js
@@ -5,7 +5,7 @@ import { PathReporter } from './PathReporter';
 export const ThrowReporter: Reporter<void> = {
   report: validation => {
     if (validation.isLeft()) {
-      throw PathReporter.report(validation).join('\n');
+      throw PathReporter.report(validation);
     }
   }
 };
