+++ test/index.js
@@ -1,17 +1,5 @@
 //@flow
-import './fp.js';
-import './Type.js';
-import './default-types.js';
-import './array.js';
-import './dictionary.js';
-import './exact.js';
-import './inexact.js';
-import './intersection.js';
-import './keyof.js';
-import './literal.js';
-import './PathReporter.js';
-import './readonlyArray.js';
-import './readonly.js';
-import './refinement.js';
-import './tuple.js';
-import './union.js';
+import './types';
+import './fp';
+import './PathReporter';
+import './ThrowReporter';
