+++ test/types/dictionary.js
@@ -1,7 +1,7 @@
 //@flow
 import * as assert from 'assert';
-import * as t from '../src/index';
-import { assertSuccess, assertFailure, assertStrictEqual, assertDeepEqual, string2, DateFromNumber } from './helpers';
+import * as t from '../../src/index';
+import { assertSuccess, assertFailure, assertStrictEqual, assertDeepEqual, string2, DateFromNumber } from '../helpers';
 
 describe('dictionary', () => {
   it('should properly typecheck values', () => {
