+++ test/types/index.js
@@ -1,0 +1,12 @@
+//@flow
+import './props';
+import './array';
+import './core';
+import './basic';
+import './dictionary';
+import './intersection';
+import './keyof';
+import './literal';
+import './refinement';
+import './tuple';
+import './union';
