+++ test/types/props/index.js
@@ -1,0 +1,4 @@
+//@flow
+import './base.js'
+import './exact.js';
+import './inexact.js';
